backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Todo App with JWT authentication and Neon PostgreSQL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.22",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.3.0",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "mypy>=1.10.0",
    "ruff>=0.6.0"
]
